AAudio: Remove TODO about deadline bug
A comment should be removed.

The comment was added when the log statements
were errors. After the comment was added, the
log statements got changed into warnings. Through
a quick bug search, these warnings seem useful
to OEMs when they encounter the warnings. Thus,
this CL simply removes the comment.

Bug: 243432373
Test: TreeHugger

Change-Id: I2153ef8dc11f67b481f297598f6e5b50e6deb2bf