libaudiohal: Make effect descriptors cache thread-safe
Due to legacy reasons, the code in the audio server assumes that
querying all effect descriptors is a cheap operation (stems from
pre-Treble times). Thus, the HIDL shim code needs to cache the
list of descriptors. It was assumed that the audio server code
queries effect descriptors under a lock, however recently it was
observed that it's not always the case (or it's not always the
same lock). To prevent races, access to the cache has been
serialized.

Bug: 243877224
Test: atest android.media.audio.cts.AudioEffectTest
Change-Id: Id5f3615074e540afca33a83d53d2b84634df89cf