Camera: Look up client name for NDK earlier
Before, the client package name, which can be empty for NDK clients,
wasn't looked up until BasicClient. Move lookup to connectHelper,
to ensure there's a name to pass to other system services such as
window manager.

Note that for shared UID processes, we may select the wrong name,
but this is not expected to cause problems for the situations where
the name is used, beyond just confusing logs.

Also fix up PID reporting for the initial camera connect logging.

Test: atest CtsCameraTestCases on oriole, plus manual testing
  with an NDK-using app to verify that the frame rate override reaches
  window manager correctly.
Bug: 239452794
Change-Id: I006a73e0c9eec99c827085a722d0edfa879bdd23