Camera: Avoid over-delaying frames in PreviewFrameSpacer
There could be 1 or 2 ms between calculating the timeToWait and
queueBuffer(). As a result currently we are over-delaying each
frame for about 1ms. This over-delay can accumulate and recover
once the maximum 10ms limit is reached or 2 frames are queued.

To compensate, wait up to until (mLastPresentTime + captureInterval - 2ms)
So that the presentation intervals don't always exceed captureInternal.

Bug: 251183098
Test: Camera CTS, vendor testing
Test: testPreviewJitterSurfaceTexture
Change-Id: I2728d0516686950141c8e6d41e041d8d6dc00c8a