Use cached invalid tracks id when checking invalid tracks.
When checking invalid tracks, audio flinger will need to call the
callback if it exists. This need to be done without the lock in case it
takes a long time. However, after unlocking, the list of tracks can
already be modified. This can cause unexpected crash. To resolve the
issue, make a copy of the invalid tracks id under lock and call the
callback outside of the lock.

Test: OboeTester test disconnect
Bug: 228582029
Change-Id: I6fe28fa097acb91f4081666bad8289eae1c62c6e