Return DISCONNECTED state when calling from API.
The disconnected state is public API and some apps relying on the state
changed to stop and restart a new stream when disconnection happens. In
android U, the disconnected state is separated from stream state in
aaudio implementation. However, the public API should still be able to
return disconnected state so that apps replying on state changed will
not be broken.

Bug: 254928976
Test: repo steps in bug
Change-Id: Ic267933a01839a22f1b39d14291f9d85757c0e0f