audio policy: fix sensor privacy listener
Current way of registering to the sensor privacy listener works
only when disabling sensors from the developer settings.
This CL also registers a listener for when the mic privacy toggle
is used.
This change is temporary until problems around app ops start/stop
notifications from audio framework are sorted out and we can rely
on app ops to silence audio when the mic privacy toggle is used.

Test: make
Change-Id: I1178a15583596cd025d5cb1e424f9fe3bf7854b4