codec2 hal: type conversion refactoring, step 2
Implement ParamTypes.h and refactor HIDL type conversion to use it.

ParamTypes.h contains template functions to be reused for
both HIDL and AIDL interfaces. The interfaces are mostly similar, and
that is why this approach works (mostly). There are cases HIDL and AIDL
interfaces are different or relying on specific HIDL/AIDL libraries,
and we use template specialization to handle those cases.

Bug: 251850069
Test: presubmit
Change-Id: I8fb24d1aaebc8fa4a34c40129715a3f146023430