Skip setting the state for paused binder thread invocations in waitUntilStateThenRelock
Because of a potential deadlock that exists between reconfigureCamera
and createInputStreams/createStreams (b/241137777), idle state from
reconfigureCamera should not notify the places waiting for idle
as that state is immediately transitioned to active. This way
reconfigureCamera will be able to complete successfully.

This will allow us to fix the deadlock by allowing us to revert
ag/10211554, as it handles the serialization problem that the revert
introduces.

Test: Camera CTS continue to pass
Bug: 241137777
Bug: 269092476

Change-Id: I9d3f13be59cf3f5f4ce695b045cd1a33e5ea1233