gav1: support 444 and 422 for 10bit (YV12/P010 paths)
This will enable 444 and 422 support for AVIF images as well.
The HAL_PIXEL_FORMAT_RGBA_1010102 path is left as a follow up.

I?10ToP010 were added in:
  7ed0473 Update external/libyuv to r1837
and upstream in version 1779.

To keep this compatible with mainline LIBYUV_VERSION checks are added
around the new code. libyuv is not a part of mainline and is currently
at libyuv version 1732.

Bug: 261112792
Test: decode images on oriole and panther devices
Test: compile pre/post libyuv update
Change-Id: I2593552a5cb9c343f7939d32803ed61e3e8f47e3