CSD: fix race condition when releasing patch
For short streams it can happen that the updateMetadata which
invalidates the active stream is sent before the patch is released. This
leads to a double calling of stopMelComputation for the same stream,
which has an effect of lowering the counter of mActiveDeviceTypes twice.

Also improved the logging for debugging.

Test: logs and dumpsys
Bug: 265928284
Change-Id: I1f1e585c974632d015d318fe616364009cefdd96
Merged-In: I1f1e585c974632d015d318fe616364009cefdd96
(cherry picked from commit 2750604b0c8e5ada62ad45b93153dfd48fe7d2d8)