Add profile and level keys to MetaDatabase
Add audio keys for profile and level to be used by audio codecs and
extractors, e.g. DTS. Reuses the values for the video profile and level
since the codec framework doesn't distinguish between video and audio
profiles so there is no need to define a separate key.

Also bringing back MEDIA_MIMETYPE_AUDIO_DTS_UHD with no profile in the
mime type to be set in MetaDatabase for the mime key (and the profile in
a separate fild). This allows for a clean mapping with the MediaFormat
SDK values (MIMETYPE_AUDIO_DTS_UHD).

Bug: 260223911
Test: atest CtsMediaExtractorTestCases
Change-Id: I986815a3e2e85414e76873c737296c1f4d053ac3