cameraservice: Add logId to correlate CameraActionEvents
Cameraservice currently triggers logging of CameraActionEvents (through
CameraServiceProxy) which are logged as independent event. For data
analysis purposes, it would be helpful to have one set of
OPEN-n*SESSION-CLOSE events correlate with each other.

To help with that, an mLogId field was added in CameraSessionStats which
will hold a string identifier. This identifier will be randomly
generated for OPEN events and reused for corresponding
ACTIVE, IDLE, and CLOSE events.

Bug: 271297241
Test: Manually tested that statsd gets the expected logId for camera
      sessions.
Change-Id: Ia26f2cd5d75bdeb15737e85d7df1a46b75993712