cameraservice: Pipe session identifier to the Camera HAL
For data analysis from HAL logs, it is helpful to get some higher level
information of the camera usage (such as the client's package name).
HALs are denied access to this higher level information at runtime.

To help data with data analysis, cameraservice will now pass a session
identifier when configuring streams. This identifier can be used to
correlate HAL logs with cameraservice logs at the time of log
consumption, but not at runtime. This identifier has no functional
purpose and is randomly generated, so does not contain any identifying
information about the high level client.

Bug: 271171714
Test: Manually verified that cameraservice passes the session identifier
      to the camera HAL.
Change-Id: Id980b11f9062369e5bca354a5113d6b5040de8be