Log additional NuPlayer information in bug reports
This change adds more logging from NuPlayer when obtaining
a bug report. NuPlayer's renderer will print whether it has occupied
the wake lock and for how much time. In addition, handlers attached
to the player's looper will also log if they are currently executing
a message (at the time of the bug report).

The above logs will help detect if the NuPlayer's wake lock is kept
acquired because the player looper is blocked on a specific handler.

Test: manual on device test, grabbed a bug-report while playing media

Change-Id: I3e91da9ea45a5fe270a0866054f695f5e844e8c7
Bug: 265531109