Declare `ConversionResult` in `android` namespace
Previously the declaration was depending on the `BACKEND_NDK`
define, and `ConversionResult` could end up in the `aidl` NS
(::aidl::ConversionResult). Having it in `android` is more
convenient as it works for nested namespaces of `android`, for
example `android::effect`.

Added alias `aidl::android::ConversionResult` for the NDK
backend case.

Bug: 205884982
Test: m
Change-Id: Ic3238144d35afb86db57d5d620f2f28ccd89021e