CSD: Remove A2DP devices for CSD calculation
Currently there is no good way to identify whether an a2dp device is a
bluetooth headphone for which the sound dose calculation can be applied
or just a bluetooth speaker (e.g.: car kit). Disabling the CSD
calculation for a2dp for now until we have a good solution for this use
case.

Test: dumpsys audio, audio_flinger
Bug: 276660720
Change-Id: I762c8f94ea6557a2aef2945a7ef1f8ac685f4577