Add recording of detailed information for freezes and judders
Add judder and freeze events which contain detailed information about
clustered groups of freeze and judder occurrences. This information is
contained in the media.metrics dumpsys for inclusion in bug reports, but
is not enabled by default and must be configured as enabled during a
debugging session or configured as enabled based on device-ids or
similar.

Bug: 234833109
Test: atest VideoRenderQualityTracker_test
Test: play YouTube videos with events programatically enabled
Change-Id: If790eb8444a160bfb7629df7160454ebe710410d