Send 'closing=true' from StreamOutHalHidl::exit
Combine sending of 'closing=true' with 'exiting=1' for legacy
HALs. Historically, the remote submix HAL was using a separate
notification because initially 'closing=true' was only sent
for the A2DP output. Unify sending of these parameters as in
the AIDL HAL interface we only use the thread exiting condition
to notify the HAL about the need to stop its output.

Bug: 73175392
Test: manual check of A2DP connection, disconnection, switching
      "Mono audio" on and off;
      manual check of screen recording with audio
Change-Id: I722e5dadbab252ccae29eedad92e568ffe664b4d