Camera: Handling out-of-order request id
For the API1 path, the request id is assumed to be in order when syncing
with the device for the request to apply. This is not always the
case, as errors can cause out-of-order and this can cause delay in the
API1 path.

This change will address this by maintaining the request id list.

Bug: 288065275
Test: Manual test with the app using API1,
atest -c -d cts/tests/camera/src/android/hardware/cts/

Change-Id: Iaacf349408c751e178173d6e03d41f1335749c27