audio flinger: fix initial mmap stream volume
The initial stream volume values received by a MMAP thread
while the mmap stream is being open by audio policy manager
were ignored because the stream type was not yet initialized (the stream
type is received from audio policy manager when getOutputForAttr
returns).

The fix consists in using the same volume/mute data structure array for
MmapPlaybackThread as for PlaybackThread in order to store all stream
volume and mute values but only use the one corresponding to the
configured mStreamType.

Bug: 284836454
Test: repro steps in bug
Change-Id: If4654c1d9ea7fd109a8ab2b9b2431db71b966243