APC: Add possible deadlock fix when creating track
When creating a new track the thread lock is being held. This can have
the effect of causing a deadlock when the OpPlayAudioMonitor is being
created since the onFirstRef method could also try to access the same
thread lock. Removing the broadcast when calling onFirstRef which is not
needed and would also avoid the deadlock

Test: atest AudioPlaybackConfigurationTest
Bug: 289885996
Change-Id: I16555608176f1b1dfa180e82346319b55c19de01