CSD: Disable the forced use of framework MEL
The vendors can decide whether to use the sound dose feature and are
responsible for ensuring that the sound dose HAL is present for modules
that support offloading. It is ok to fall back to the internal framework
MEL computation when a CSD HAL is not implemented. This will however not
guarantee a certification with the EN50332-3 and IEC62368-1 3rd edition
regulation.

Test: manual CSD enablement
Bug: 287011781
Change-Id: If6a6f47a408020011fcb7a4af2dbd0c0121b1c86