AIDL: Consider both inputs and output of r_submix to be "virtual"
Since the AIDL implementation of the remote submix does support
"connect external device" operation, the framework uses it to assign
a device address to a remote submix pipe. Thus, we should treat
both inputs and outputs as "virtual" connections, that is, they
need to be connected explicitly, however they don't require
a physical device.

Bug: 286914845
Bug: 294976817
Test: live caption scenario
Change-Id: I0e5d13e03ebf0d466080f83cc9373906e9affa36