Add a boolean for peek enabled
Peek state is always changed to kEnabledNoBuffer after start()
and flush(), if it was not in legacy mode. This requires app
to always disable peek after start or flush if it wants to.

The fix is Adding a boolean to indicate if peek is enabled and
set the state accordingly in start() and flush().

Test: atest DecodeOnlyTest
Bug: 287456023
Change-Id: I5b4f0ea1c1740ba94f59eaf869bd31ae049127c0