AIDL: Fix test after aosp/2704653
That change updated the connection for the "remote submix in"
device type, however the test was not updated, resulting in
a failure.

The need to modify the test was overlooked because
the TEST_MAPPING file for the library was wrong, although
syntacsically correct. Fix the mapping file to prevent
future regressions.

Bug: 286914845
Bug: 294976817
Bug: 296664539
Test: audio_aidl_conversion_tests
Change-Id: I3bcfab9ca45c19d487deb7709d32f43beee8c789
Merged-In: I3bcfab9ca45c19d487deb7709d32f43beee8c789