DO NOT MERGE Camera: Sync request thread exit with request clear flag
'mRequestClearing' will be set before the request thread is
able to exit. Depending on timing, the request thread can clear
the flag and wait for the request signal one extra
iteration before it is able to quit.
Avoid the extra wait by requesting exit and enabling the request
clear flag synchronously.

Bug: 297497171
Test: Camera CTS
Change-Id: I7856e633f897e245ca2df7155c6b18c1e6efd81f