libaudiohal@aidl: Fix `setPortConfigFromConfig` for unspecified values
AudioPortConfig type expects that unspecified parameters shall be
left unset: they are @nullable to match the legacy approach with
flags which was used in struct audio_port_config. However,
`setPortConfigFromConfig` was mistakingly setting parameters even
when they had default values in the source AudioConfig. This was
confusing the implementation of `IModule::setAudioPortConfig`.

Bug: 298175108
Test: atest audiosystem_tests
Change-Id: Ie352bb3874a9d0a357a30e766ebb6ba14ab6e0c9