Camera: Add support test pattern session parameter
If test pattern related keys are in the session parameter list then the
null capture request could be sent to the HAL after reconfiguration when
camera privacy option is changed.

This patch will add support for the test pattern when the keys are in
the session parameter list.

Bug: 316170074
Test: Manual test with the device
Change-Id: If7d3de517d1061a8f2997fae624144fe90e9ce1f
Merged-In: If7d3de517d1061a8f2997fae624144fe90e9ce1f