audiopolicy: fix audioflinger patch release while using SwBridge
Test: make

When using a SwBridge, AudioFlinger creates a new patch to reach the
sink device from the SwOutput.
Upon release, the patch is released, hence the associated mixer thread
has released the patch.
Previously, setOutputDevices was forced on the SwOutput to recreate a
patch.
However, it consumes times, hence leading to for example certification
issue of CarPlay (call shall be routed within less than 100ms while in
FM)

Change-Id: I84209b8e69d202e5754946a341f8bb1d89b7d5a6
Signed-off-by: François Gaffie <francois.gaffie@renault.com>