Update and clean up audiorouting_tests
The important change is in TestPerformanceMode to not require
the mix port serving a FAST track to have a FAST flag. This is
because a FastMixer potentially can be created on any port
which satisfies certain buffer size conditions.

Remove the code which reads the XML APM file. The test does not
use it (and shouldn't as the XML file may not exist with the AIDL
HAL).

Add tracing of test invocations to correspond logcat with
test start/completion.

Also, replace hand-written dumping function with invocations
of `toString` functions from AIDL parcelables.

Bug: 302036943
Test: atest AudioTrackTest
Change-Id: Ia81cf429af92b8ac9582abc7cf270b1ccb96019c