audio AIDL: Consider `BUS` devices as attached
After analyzing Auto APM config files and checking with
the Auto team, it has turned out that `BUS` devices are
considered to be permanently attached. In the AIDL schema
that means, they must have an empty connection type.

A solution which seems to work well is to map AIDL types
`IN_|OUT_DEVICE` with no connection to legacy BUS devices.

Bug: 303926380
Test: atest audio_aidl_ndk_conversion_tests
Change-Id: Id98d632f47f4b4498fac5f92e092bc7d7485482e