Remove duplicate watchdog monitoring for disconnect/close
- Since the watchdog monitoring for Camera3Device::close is nested within another watchdog call for Camera2ClientBase::disconnectImpl,
there is potential for a deadlock to occur within the watchdog when one
of the nested calls aborts. While this situation is not entirely clear,
it is safer to remove the redundancy and avoid the nested call.
- We retain the monitoring of 'close' and not 'disconnectImpl’ because
'close' is a few function calls closer to the HAL and the watchdogs intention is to monitor deadlocks in the HAL without interfering with existing error
handling.

Test: Camera CTS Test
Bug: 302168348
Change-Id: I56cde7278009ef8c72b74469f29d1914f663adb8